+++ changelog.rst
@@ -2,6 +2,12 @@ ChangeLog
 =========
 
 
+`5.2.0 (unreleased) <https://github.com/kdeldycke/mail-deduplicate/compare/v5.2.0...develop>`_
+----------------------------------------------------------------------------------------------
+
+.. note:: This version is not yet released and is under active development.
+
+
 `5.1.0 (2020-10-06) <https://github.com/kdeldycke/mail-deduplicate/compare/v5.0.0...v5.1.0>`_
 ---------------------------------------------------------------------------------------------